+++ src/server/sv_game.c
@@ -1043,7 +1043,7 @@ intptr_t SV_GameSystemCalls(intptr_t *args)
 		return SV_BinaryMessageStatus(args[1]);
 
 	default:
-		Com_Error(ERR_DROP, "Bad game system trap: %i", args[0]);
+		Com_Error(ERR_DROP, "Bad game system trap: %ld", (long int) args[0]);
 	}
 	return -1;
 }
